requirements.txt
python-telegram-bot==20.0
requests==2.26.0
beautifulsoup4==4.10.0
apscheduler==3.8.1
pytz==2021.1